Update in ds_helper to support import to Pinnacle
I have updated the function add_refd_frame_of_ref_sequence in ds_helper to support the import of files into the treatment planning system Pinnacle. There was a note in the code that the previous version should be checked. In Pinnacle, the selection of and random UID for Frame of Reference was not accepted during the import.